+++ app/src/main/kotlin/com/atherton/upnext/presentation/navigation/Navigator.kt
@@ -3,9 +3,9 @@ package com.atherton.upnext.presentation.navigation
 interface Navigator {
 
     fun showSearchScreen()
-    fun showMovieDetailScreen(movieId: Int)
-    fun showTvShowDetailScreen(tvShowId: Int)
-    fun showPersonDetailScreen(personId: Int)
+    fun showMovieDetailScreen(movieId: Long)
+    fun showTvShowDetailScreen(tvShowId: Long)
+    fun showPersonDetailScreen(personId: Long)
     fun playYoutubeVideo(videoKey: String)
     fun showSettingsScreen()
     fun showLicensesScreen()
